Extract Version 33.03 from BBMap_33.04_java7.tar.gz

+++ current/jgi/BBDuk2.java
@@ -1373,8 +1373,8 @@ public class BBDuk2 {
 		private long addToMap(Read r, int skip){
 			skip=Tools.max(minSkip, Tools.min(maxSkip, skip));
 			final byte[] bases=r.bases;
-			final long shift=2*k;
-			final long shift2=shift-2;
+			final int shift=2*k;
+			final int shift2=shift-2;
 			final long mask=~((-1L)<<shift);
 			final long kmask=kMasks[k];
 			long kmer=0;
@@ -1967,8 +1967,8 @@ public class BBDuk2 {
 		final byte[] bases=r.bases, quals=r.quality;
 		final int minlen=k-1;
 		final int minlen2=(maskMiddle ? k/2 : k);
-		final long shift=2*k;
-		final long shift2=shift-2;
+		final int shift=2*k;
+		final int shift2=shift-2;
 		final long mask=~((-1L)<<shift);
 		final long kmask=kMasks[k];
 		long kmer=0;
@@ -2119,8 +2119,8 @@ public class BBDuk2 {
 		final byte[] bases=r.bases;
 		final int minlen=k-1;
 		final int minlen2=(maskMiddle ? k/2 : k);
-		final long shift=2*k;
-		final long shift2=shift-2;
+		final int shift=2*k;
+		final int shift2=shift-2;
 		final long mask=~((-1L)<<shift);
 		final long kmask=kMasks[k];
 		long kmer=0;
@@ -2173,8 +2173,8 @@ public class BBDuk2 {
 		final byte[] bases=r.bases;
 		final int minlen=k-1;
 		final int minlen2=(maskMiddle ? k/2 : k);
-		final long shift=2*k;
-		final long shift2=shift-2;
+		final int shift=2*k;
+		final int shift2=shift-2;
 		final long mask=~((-1L)<<shift);
 		final long kmask=kMasks[k];
 		long kmer=0;
@@ -2236,8 +2236,8 @@ public class BBDuk2 {
 		final byte[] bases=r.bases;
 		final int minlen=k-1;
 		final int minlen2=(maskMiddle ? k/2 : k);
-		final long shift=2*k;
-		final long shift2=shift-2;
+		final int shift=2*k;
+		final int shift2=shift-2;
 		final long mask=~((-1L)<<shift);
 		final long kmask=kMasks[k];
 		long kmer=0;
